server/requirements.txt
# Core FastAPI and web framework dependencies
fastapi==0.115.2
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1

# LangChain and LangGraph dependencies for multi-agent workflow
langchain==0.3.26
langchain-core==0.3.67
langchain-community==0.3.27
langgraph==0.3.5

# Vector database and retrieval dependencies
faiss-cpu==1.8.0

# Embeddings and semantic search
sentence-transformers==3.1.1
numpy==1.26.4


# Data processing and utilities
pandas==2.2.3

# Authentication and security dependencies
PyJWT==2.9.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Configuration and data validation
pydantic==2.9.2
pydantic-settings==2.5.2

# HTTP client utilities
httpx==0.27.2

# Additional utility dependencies
typing-extensions==4.12.2
python-dateutil==2.9.0.post0

# Text processing dependencies
lxml==5.3.0

# Optional dependencies for extended functionality
google-api-python-client==2.141.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1

# Optional dependencies for periodic RSS ingestion tasks
# apscheduler==3.10.4

# Optional dependencies for additional text processing
# unstructured==0.15.11

# Explicit version constraints to maintain compatibility
accelerate==1.0.1
torch==2.4.1
transformers==4.44.2

loguru==0.7.3
pyttsx3==2.99